Reduce memory footprint a bit
Some small strings and variable resizes.  Seems to be a 24 byte memory leak
on every MWTT stop/start.

+++ ntp.cpp
@@ -31,7 +31,7 @@ static WiFiUDP ntpUDP;
 static time_t GetNTPTime();
 static void SendNTPPacket(IPAddress &address, byte *packetBuffer);
 
-static int syncInterval = 600;
+static short syncInterval = 600;
 
 void StartNTP()
 {
